+++ test/test_full.py
@@ -2,10 +2,10 @@ import json
 import pytest
 from gmft.pdf_bindings.bindings_pdfium import PyPDFium2Document
 from gmft.table_detection import TableDetector
-from gmft.table_function import AutoTableFormatter
+from gmft import AutoTableFormatter
 
 
-@pytest.fixture
+@pytest.fixture(scope="session")
 def doc_tiny():
     doc = PyPDFium2Document("test/samples/tiny.pdf")
     yield doc