+++ test/test_pdfium.py
@@ -7,7 +7,7 @@ import gmft
 from gmft.pdf_bindings import PyPDFium2Document
 
 
-@pytest.fixture
+@pytest.fixture(scope="session")
 def doc_tiny():
     doc = PyPDFium2Document("test/samples/tiny.pdf")
     yield doc