+++ test/test_serial.py
@@ -11,14 +11,14 @@ from gmft.pdf_bindings.bindings_pdfium import PyPDFium2Document
 from gmft.table_detection import RotatedCroppedTable
 from gmft.table_function import TATRFormattedTable
 
-@pytest.fixture
+@pytest.fixture(scope="session")
 def doc_tiny():
     doc = PyPDFium2Document("test/samples/tiny.pdf")
     yield doc
     # cleanup
     doc.close()
 
-@pytest.fixture
+@pytest.fixture(scope="session")
 def doc_9():
     doc = PyPDFium2Document("test/samples/9.pdf")
     yield doc