+++ CHANGELOG.md
@@ -1,11 +1,12 @@
 ## v0.2.0-pre
 
-- Return visualized images as PIL image
+- Return visualized images as PIL image; allow padding or margin around visualized
 
 Several tweaks to formatting algorithm that may result in different outputs compared to prior versions.
 - Automatically drop rows whose only non-null values is the "is_projecting_row" column
 - Fill in gaps between table rows, so hopefully no text is skipped
 - ^ TODO: in that case, try the large table assumption
+- Non-maxima suppression, as seen in inference.py
 - Widen out the rows to same length, as seen in inference.py
 - Several tweaks to conditions, parameters, heuristics